Moving Unused Classes
Added Crossbolter effect
