+++ src/main/java/me/shadowalzazel/mcodyssey/common/world_events/OminousStructureEvent.kt
@@ -1,0 +1,4 @@
+package me.shadowalzazel.mcodyssey.common.world_events
+
+class OminousStructureEvent {
+}
