+++ src/main/java/me/shadowalzazel/mcodyssey/common/world_events/daily_events/DayEvents.kt
@@ -1,0 +1,12 @@
+package me.shadowalzazel.mcodyssey.common.world_events.daily_events
+
+import me.shadowalzazel.mcodyssey.common.world_events.daily_events.day_events.SlimeShower
+
+object DayEvents {
+
+    val SLIME_SHOWER = SlimeShower()
+
+
+    val eventList = listOf(SLIME_SHOWER)
+
+}
