+++ src/main/java/me/shadowalzazel/mcodyssey/common/world_events/daily_events/night_events/AbyssalNight.kt
@@ -1,0 +1,4 @@
+package me.shadowalzazel.mcodyssey.common.world_events.daily_events.night_events
+
+class AbyssalNight {
+}
