+++ src/main/java/me/shadowalzazel/mcodyssey/common/world_events/daily_events/night_events/BloodMoon.kt
@@ -1,0 +1,18 @@
+package me.shadowalzazel.mcodyssey.common.world_events.daily_events.night_events
+
+import me.shadowalzazel.mcodyssey.common.world_events.daily_events.ActivationTime
+import me.shadowalzazel.mcodyssey.common.world_events.daily_events.DailyWorldEvent
+import me.shadowalzazel.mcodyssey.common.world_events.utility.EntityConditions
+
+class BloodMoon : DailyWorldEvent(
+    "blood_moon",
+    4,
+    4,
+    ActivationTime.Night,
+    listOf(EntityConditions.AlwaysTrue),
+) {
+
+
+
+
+}
