+++ src/main/java/me/shadowalzazel/mcodyssey/common/world_events/ominous_events/OminousEvent.kt
@@ -1,0 +1,9 @@
+package me.shadowalzazel.mcodyssey.common.world_events.ominous_events
+
+// Called when drinking bad omen inside structure
+class OminousEvent {
+
+
+    // CALLED ON ITEM CONSUME EVENT
+
+}
