+++ src/main/java/me/shadowalzazel/mcodyssey/unused/bosses/the_ambassador/AmbassadorMessage.kt
@@ -1,0 +1,26 @@
+package me.shadowalzazel.mcodyssey.unused.bosses.the_ambassador
+
+enum class AmbassadorMessage(val text: String) {
+
+    MAX_LIKENESS(", I am sorry. I can not receive any more pleasantries from you."),
+    WAIT_FOR_GIFT_COOLDOWN(", please wait. Still... thinking..."),
+    SKY_BOMBARD_ATTACK("I like fireworks, the heavens and gifts. But not you."),
+    GRAVITY_LAUNCH_ATTACK("The universe is unfathomable. Your are just mundane..."),
+    HIJACK_ATTACK("It appears your friends are actually your foes..."),
+    FALLING_SINGULARITY_ATTACK("The point is that it's super massive..."),
+    ELYTRA_PULL_BACK_ATTACK("You dare use such puny voidflight apparatus."),
+    PATIENCE_MESSAGE_1(" threatens your safety..."),
+    PATIENCE_MESSAGE_2(" is testing my patience!"),
+    PATIENCE_MESSAGE_3("The lack of intelligence on this test world is apparent."),
+    PATIENCE_MESSAGE_4("... I expected more from you..."),
+    PATIENCE_MESSAGE_5("... Does your integrity not amount to anything?!"),
+    PATIENCE_MESSAGE_6(", that is not an appropriate way to introduce yourself, though what is expected from such a lowlife..."),
+    PATIENCE_MESSAGE_7("'s insolence has ended my graciousness. Fear me, if you dare."),
+    PATIENCE_MESSAGE_8("... Time to teach the lowlifes basic manners..."),
+    PATIENCE_MESSAGE_9("Who is doing that?!"),
+    REJECT_ENCHANTS("The Arcane is not something to give..."),
+    NICE_GIFTS("Your gifts are enjoyable."),
+    SCULK_GIFT("The Sculk has corrupted this forsaken place... Its eradication is inevitable"),
+    BAD_GIFTS("This is not a gift, but left over debris, suitable for fodder.")
+
+}
