--- src/main/java/me/shadowalzazel/mcodyssey/world_events/FestiveEvents.kt
@@ -1,7 +1,0 @@
-package me.shadowalzazel.mcodyssey.world_events
-
-enum class FestiveEvents { // Events that target villages and bases
-
-
-
-}
