--- src/main/java/me/shadowalzazel/mcodyssey/world_events/OminousEvents.kt
@@ -1,18 +1,0 @@
-package me.shadowalzazel.mcodyssey.world_events
-
-enum class OminousEvents { // Triggered by Bad Omen
-
-
-    OMINOUS_LEGION, // A Horde of pillagers appears at a beacon 'BEACON LVL + OMEN LVL'
-    OMINOUS_INCURSION, // A Ruined Portal Becomes an active portal and piglins invade
-    OMINOUS_MINESHAFT, // A Mineshaft becomes ghastly and destructive mobs appear
-    // Triggered at night
-
-
-}
-
-// Create enum for triggers
-// Night
-// Enter Base
-// Enter Structure
-// Enter Biome
