--- src/main/java/me/shadowalzazel/mcodyssey/world_events/OminousStructureEvent.kt
@@ -1,4 +1,0 @@
-package me.shadowalzazel.mcodyssey.world_events
-
-class OminousStructureEvent {
-}
