--- src/main/java/me/shadowalzazel/mcodyssey/world_events/daily_events/DayEvents.kt
@@ -1,12 +1,0 @@
-package me.shadowalzazel.mcodyssey.world_events.daily_events
-
-import me.shadowalzazel.mcodyssey.world_events.daily_events.day_events.SlimeShower
-
-object DayEvents {
-
-    val SLIME_SHOWER = SlimeShower()
-
-
-    val eventList = listOf(SLIME_SHOWER)
-
-}
