--- src/main/java/me/shadowalzazel/mcodyssey/world_events/daily_events/night_events/AbyssalNight.kt
@@ -1,4 +1,0 @@
-package me.shadowalzazel.mcodyssey.world_events.daily_events.night_events
-
-class AbyssalNight {
-}
